requirements.txt
langchain_openai
langchain_core
langchain_community
streamlit
python-dotenv
fastapi
langserve
uvicorn
google-generativeai
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
sentence-transformers
langchain
